% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_pd.R
\name{play_pd}
\alias{play_pd}
\title{Play_pd}
\usage{
play_pd(starttheme = "princesses")
}
\arguments{
\item{start_theme}{This argument is not working yet!
App will start at default theme "princesses".
(String character used to make the game start at the given theme.
Either the full theme name or the abbreviation (i.e. "prin", "dino") is
valid. Default is princesses.)}
}
\value{
No values are returned to R.
}
\description{
Launches the Shiny interface to play Princesses and Dragons
The app is developed for toddlers and pre-school age children (ca. 2-6 yr).
They can play two games: Memory and Stop!. They can also switch between two
themes: "princesses" or "dinosaurs".
}
\examples{
\dontrun{
play_pd("dino")
}

}
